Remove v2 file
